requirements.txt
# requirements.txt

# The core web framework for building the API
fastapi==0.111.0

# The ASGI server needed to run the FastAPI application.
# The [standard] option includes helpful extras for development.
uvicorn[standard]==0.29.0

# The official Python client library for the OpenAI API
openai==1.35.3

# The HTTP client library used by OpenAI; pinning the version prevents conflicts.
httpx==0.27.0

# Reads key-value pairs from a .env file and sets them as environment variables
python-dotenv==1.0.1